Modified sleep to cast the calculated millisecond value to a DWORD.

+++ source/library/ash.system.windows.clock.h
@@ -95,7 +95,7 @@ namespace Ash
 
 				static inline void sleep(Nanosecond duration)
 				{
-					::Sleep(duration / 1000000ULL);
+					::Sleep(DWORD(duration / 1000000ULL));
 				}
 
 			protected:
